Added include/v3x and src/v3x files. They are just a copy of what's in include/v1x and src/v1x, but using the v3x namespace. We don't want to create a new set of classes for the time being. We will be working on a small subset of cQasm 1.0.
Added code to deal with those files in src/CMakeLists.txt.
I am trying to add the generation of the ANTLR parser as a custom target, the same we are doing for the flex/bison parser.

Created a generate_antlr_parser.py.
I am trying to generate the ANTLR parser in a portable way, from different OS.