Fix Python API
It was broken: the "automated" typemap for std::vector<std::string> was
somehow not working, and returning an opaque SWIG object which the
Python wrapper couldn't access.

It turned out that most of the typemaps in libQasm.i were not needed, so
this file is reduced.

I've also added a test for the Python API.