Added include/version.hpp to define libqasm version in a single place. Then made this definition available to Python code through get_version method at version.py. And used this method from setup.py.
This is very similar to what is done at the moment for the QX simulator.
The main difference is that I've put the implementation of the get_version method in a version.py file.
I've done this because I think, in the future, it may be needed to access this get_version method from conanfile.py.