Implemented initializations for axis type.
An axis variable is considered to always have size 3.

For cQasm 1.x, axis type was implemented as an enum taking 3 possible values: X, Y, or Z.
For cQasm 3.0, axis type is implemented as an array of 3 real values, each corresponding to vectors on the X, Y, and Z axes.

Boolean, integer, and real arrays promote to axis (an array of 3 real values).