Reimplement AnalyzeTreeGenAstVisitor as a subclass of ast::Visitor<std::any>.
There is even more boilerplate code now.
I think this is due to the fact that visit methods use references as parameters.
That forbids polymorphism of nodes to kick in.
Instead, as_<node_type> has to be used to visit the appropriate nodes.

Also, callers to visitor methods have to std::any_cast the return values, because visitor methods return std::any.

On top of that, some methods return an empty std::any.

On the plus side, visit_index_list, visit_index_item, and visit_index_range do not need the extra size argument.
The out of range checks are done at visit_index.

visit_index_list has been fixed to not discard previous indices when traversing an index range.

Some throw statements have been fixed to not include an AST node.
This is not needed if the throw statement is caught later and a context added to the error.