WIP.
Add 'inv', 'pow', and 'ctrl' keywords.
An instruction can be a measure, a reset, or a gate.
A gate can be a modified gate or just a gate.

Update lexer.
Update parser.
Update syntactic AST.
Update semantic AST.
Update BuildTreeGenAstVisitor.
Update AnalyzeTreeGenAstVisitor.

TODO: starting with the semantic AST, the code is just wrong.